[Update] delete file method

+++ server/utils/file.js
@@ -1,8 +1,13 @@
 const fs = require("fs");
 
 const deleteFile = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) throw err;
-  });
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (err) => {
+      if (err) throw err;
+    });
+  } else {
+    console.log(`File does not exist: ${filePath}`);
+  }
 };
+
 exports.deleteFile = deleteFile;
